chore: parse attributes from the gemspec
Attributes look like this:

```
spec.name = 'example'
```

After calling Gemspec#call, the Gemspec#attributes attribute will be populated
with an array of AttributeNode objects which each refer to an attribute and a
node. The attribute has attributes name and value. The node is a
Parser::AST::Node object which represents the attribute in the gemspec's AST.